+++ runtime/datetime/datetime.h
@@ -4,10 +4,10 @@
 
 #pragma once
 
+#include "runtime/datetime/timelib_wrapper.h"
 #include "runtime/dummy-visitor-methods.h"
 #include "runtime/kphp_core.h"
 #include "runtime/refcountable_php_classes.h"
-#include "runtime/timelib_wrapper.h"
 
 struct C$DateTimeZone : public refcountable_php_classes<C$DateTimeZone>, private DummyVisitorMethods {
   using DummyVisitorMethods::accept;
