+++ runtime/datetime/timelib_wrapper.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "kphp_core.h"
+#include "runtime/kphp_core.h"
 
 // php_timelib wraps the https://github.com/derickr/timelib library
 // which is used in PHP to implement several datetime lib functions.
