+++ runtime/interface.cpp
@@ -25,7 +25,8 @@
 #include "runtime/confdata-functions.h"
 #include "runtime/critical_section.h"
 #include "runtime/curl.h"
-#include "runtime/datetime.h"
+#include "runtime/datetime/datetime_functions.h"
+#include "runtime/datetime/timelib_wrapper.h"
 #include "runtime/json-processor-utils.h"
 #include "runtime/exception.h"
 #include "runtime/files.h"
@@ -50,7 +51,6 @@
 #include "runtime/udp.h"
 #include "runtime/url.h"
 #include "runtime/zlib.h"
-#include "runtime/timelib_wrapper.h"
 #include "server/database-drivers/adaptor.h"
 #include "server/job-workers/job-message.h"
 #include "server/json-logger.h"
