+++ runtime/misc.cpp
@@ -12,7 +12,7 @@
 #include <unistd.h>
 
 #include "runtime/critical_section.h"
-#include "runtime/datetime.h"
+#include "runtime/datetime/datetime_functions.h"
 #include "runtime/exception.h"
 #include "runtime/files.h"
 #include "runtime/interface.h"
