+++ runtime/udp.cpp
@@ -11,7 +11,7 @@
 #include "common/resolver.h"
 
 #include "runtime/critical_section.h"
-#include "runtime/datetime.h"
+#include "runtime/datetime/datetime_functions.h"
 #include "runtime/net_events.h"
 #include "runtime/streams.h"
 #include "runtime/string_functions.h"
